% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insee_auth.R
\name{insee_deauth}
\alias{insee_deauth}
\title{Suspend authentication}
\usage{
insee_deauth(verbose = TRUE)
}
\arguments{
\item{verbose}{logical; do you want informative messages?}
}
\value{
NULL, invisibly.
}
\description{
Suspend access to an application.
}
