% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\name{insee_scopes}
\alias{insee_scopes}
\title{Addresses of the Insee APIs}
\usage{
insee_scopes(
  api = c("Sirene V3", "Nomenclatures v1"),
  insee_url = getOption("apinsee.url")
)
}
\arguments{
\item{api}{Un vecteur de chaînes de caractères dont chaque élément comprend
le nom d'une API. La correspondance partielle est acceptée.}

\item{insee_url}{Adresse du site fournissant les API. Ce paramètre n'est
utile que pour un usage interne à l'Insee, voir la section "Utilisation
interne à l'Insee".}
}
\value{
Un vecteur de chaînes de caractères comprenant les adresses des API
de l'Insee.
}
\description{
Cette fonction renvoie les adresses des API de l'Insee.
}
\section{Utilisation interne à l'Insee}{

Dans le cas d'utilisation standard, la valeur de l'option \code{apinsee.url} est
\code{"https://api.insee.fr/"}. Il n'est pas utile de la modifier sauf pour les
agents de l'Insee qui souhaiteraient utiliser les plateformes de test, de
recette ou de pré-production. Dans ce cas, afin de rendre les programmes
portables d'un environnement à un autre.il est préférable de déclarer
l'adresse en modifiant l'option \code{apinsee.url} :
\code{options(apinsee.url = "<URL>")}.
}

\examples{
insee_scopes()
insee_scopes("Sirene")
insee_scopes("Nomenclatures")
insee_scopes(c("Sirene", "Nomenclatures"))
}
\keyword{internal}
