% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\encoding{UTF-8}
\name{insee_token}
\alias{insee_token}
\title{Generate a valid token for an Insee application}
\usage{
insee_token(app, cache = getOption("httr_oauth_cache"),
  config_init = list(), credentials = NULL, validity_period = 86400,
  insee_url = getOption("apinsee.url"))
}
\arguments{
\item{app}{An OAuth consumer application, created by
\code{\link[=oauth_app]{oauth_app()}}}

\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics.
A string means use the specified path as the cache file.}

\item{config_init}{Additional configuration settings sent to
\code{\link[=POST]{POST()}}, e.g. \code{\link[=user_agent]{user_agent()}}.}

\item{credentials}{Advanced use only: allows you to completely customise
token generation.}

\item{validity_period}{A positive integer; token validity period in seconds.}

\item{insee_url}{Adresse du site fournissant les API. Ce paramètre n'est
utile que pour un usage interne à l'Insee, voir la section "Utilisation
interne à l'Insee".}
}
\value{
Un objet de classe \link{TokenInsee}.
}
\description{
Cette fonction génère un jeton d'accès valide à une application créée sur le
site \href{https://api.insee.fr/}{api.insee.fr}.
}
\section{Utilisation interne à l'Insee}{

Dans le cas d'utilisation standard, la valeur de l'option \code{apinsee.url} est
\code{"https://api.insee.fr/"}. Il n'est pas utile de la modifier sauf pour les
agents de l'Insee qui souhaiteraient utiliser les plateformes de test, de
recette ou de pré-production. Dans ce cas, afin de rendre les programmes
portables d'un environnement à un autre.il est préférable de déclarer
l'adresse en modifiant l'option \code{apinsee.url} :
\code{options(apinsee.url = "<URL>")}.
}

\keyword{internal}
