% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insee_auth.R
\encoding{UTF-8}
\name{insee_auth}
\alias{insee_auth}
\title{Authenticate to an Insee application}
\usage{
insee_auth(new_auth = FALSE, appname = "DefaultApplication",
  key = Sys.getenv("INSEE_APP_KEY"),
  secret = Sys.getenv("INSEE_APP_SECRET"), validity_period = 86400,
  cache = FALSE, verbose = TRUE,
  insee_url = getOption("apinsee.url"))
}
\arguments{
\item{new_auth}{booléen, valeur par défaut : \code{FALSE}. Passer \code{TRUE} si vous
souhaitez révoquer le jeton d'accès et vous authentifier à nouveau.}

\item{appname}{nom de l'application.}

\item{key, secret}{clef et secret du consommateur.}

\item{validity_period}{entier, durée de validité du jeton d'accès en secondes.}

\item{cache}{booléen indiquant si \code{apinsee} doit sauvegarder les jetons
d'accès dans un fichier cache, par défaut \code{.httr-oauth}.}

\item{verbose}{booléen; souhaitez-vous des messages d'information ?}

\item{insee_url}{Adresse du site fournissant les API. Ce paramètre n'est
utile que pour un usage interne à l'Insee, voir \link{insee_endpoint}.}
}
\value{
Un objet représentant un token pouvant être passé en valeur du
paramètre \code{token} de la fonction \link[httr:config]{httr::config()}.
}
\description{
Cette fonction permet de s'authentifier via une application créée sur
\href{https://api.insee.fr/}{api.insee.fr} : elle renvoie un jeton d'accès
valide qui peut ensuite être passé en valeur du paramètre \code{token} de
la fonction \link[httr:config]{httr::config()}.
}
\details{
Le couple clef du consommateur/secret du consommateur de l'application
devant rester secret, il est vivement recommandé de ne jamais l'inclure
dans un programme. La pratique recommandée est d'utiliser des variables
d'environnement pour stocker ces informations. La fonction \code{insee_auth()}
utilise deux variables d'environnement nommées \code{INSEE_APP_KEY} et
\code{INSEE_APP_SECRET}. Une fois renseignées, l'authentification s'effectue par
un simple appel à la fonction \code{insee_auth()}.

Les variables d'environnement peuvent être déclarées dans le fichier
\code{\link[base:Startup]{.Renviron}}. La modification de ce fichier peut s'effectuer
facilement grâce la fonction \link[usethis:edit]{usethis::edit_r_environ()}.
}
\section{Utilisation interne à l'Insee}{

Les agents de l'Insee qui souhaiteraient accéder aux plateformes de test, de
recette ou de pré-production doivent modifier en conséquence l'adresse
d'accès. Pour ce faire, il est recommandé de modifier la valeur de l'option
\code{apinsee.url}. Les programmes seront ainsi facilement portables d'un
environnement à un autre.
}

\examples{
# Modify the following option to access to a different url
# options(apinsee.url = "https://api.insee.fr/")

library(apinsee)
library(httr)

# Set the environment variables INSEE_APP_KEY and INSEE_APP_SECRET in the .Renviron file
if (all(nzchar(Sys.getenv(c("INSEE_APP_KEY", "INSEE_APP_SECRET"))))) {
  # retrieve the token
  token <- insee_auth()
  # use the token
  set_config(config(token = token))
}
}
\seealso{
\link{insee_deauth}, \link{TokenInsee}, \link[httr:config]{httr::config}.
}
