% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insee_auth.R
\name{insee_auth}
\alias{insee_auth}
\title{Authenticate to an Insee application}
\usage{
insee_auth(token = NULL, new_app = FALSE,
  appname = "DefaultApplication", key = Sys.getenv("INSEE_API_KEY"),
  secret = Sys.getenv("INSEE_API_SECRET"),
  cache = getOption("httr_oauth_cache"), verbose = TRUE)
}
\arguments{
\item{token}{optional; an actual token object or the path to a valid token
stored as an \code{.rds} file.}

\item{new_app}{logical, defaults to \code{FALSE}. Set to \code{TRUE} if you
want to wipe the slate clean and re-authenticate with the same or different
application. This disables the \code{.httr-oauth} file in current
working directory.}

\item{appname}{application name.}

\item{key, secret}{consumer key and secret of the application.}

\item{cache}{logical indicating if \code{apinsee} should cache
credentials in the default cache file \code{.httr-oauth}.}

\item{verbose}{print message.}
}
\value{
A token.
}
\description{
Authenticate to an Insee application
}
