% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insee_auth.R
\name{insee_auth}
\alias{insee_auth}
\title{Authenticate to an Insee application}
\usage{
insee_auth(new_auth = FALSE, appname = "DefaultApplication",
  key = Sys.getenv("INSEE_APP_KEY"),
  secret = Sys.getenv("INSEE_APP_SECRET"), validity_period = 86400,
  cache = FALSE, verbose = TRUE)
}
\arguments{
\item{new_auth}{logical, defaults to \code{FALSE}. Set to \code{TRUE} if you
want to wipe the slate clean and re-authenticate with the same
application. This disables the \code{.httr-oauth} file in current
working directory.}

\item{appname}{application name.}

\item{key, secret}{consumer key and secret of the application.}

\item{validity_period}{integer, length of the validity period in seconds.}

\item{cache}{logical indicating if \code{apinsee} should cache
credentials in the default cache file \code{.httr-oauth}.}

\item{verbose}{logical; do you want informative messages?}
}
\value{
A token.
}
\description{
Authenticate to an Insee application
}
