% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insee_auth.R
\name{insee_deauth}
\alias{insee_deauth}
\title{Suspend authentication}
\usage{
insee_deauth(clear_cache = TRUE, verbose = TRUE)
}
\arguments{
\item{clear_cache}{logical indicating whether to disable the
\code{.httr-oauth} file in working directory, if such exists, by renaming
to \code{.httr-oauth-SUSPENDED}}

\item{verbose}{logical; do you want informative messages?}
}
\value{
NULL, invisibly.
}
\description{
Suspend access to an application.
}
