% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\encoding{UTF-8}
\name{insee_endpoint}
\alias{insee_endpoint}
\title{Get Insee endpoint}
\usage{
insee_endpoint()
}
\value{
Un objet de classe \link[httr:oauth_endpoint]{oauth_endpoint}.
}
\description{
Cette fonction renvoie les points d'accès définis par
\href{https://api.insee.fr}{api.insee.fr}.
}
\examples{
insee_endpoint()
}
\keyword{internal}
