% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\encoding{UTF-8}
\name{insee_endpoint}
\alias{insee_endpoint}
\title{Get Insee endpoint}
\usage{
insee_endpoint()
}
\value{
Un objet de classe \code{oauth_endpoint}.
}
\description{
Cette fonction renvoie le point d'accès aux API de l'Insee.
}
\examples{
insee_endpoint()
}
