% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\encoding{UTF-8}
\name{insee_token}
\alias{insee_token}
\title{Generate a token for an Insee application}
\usage{
insee_token(app, cache = getOption("httr_oauth_cache"),
  config_init = list(), credentials = NULL, validity_period = 86400,
  insee_url = getOption("apinsee.url"))
}
\arguments{
\item{app}{An OAuth consumer application, created by
\code{\link[=oauth_app]{oauth_app()}}}

\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}, \code{FALSE} means
don't cache, and \code{NA} means to guess using some sensible heuristics.
A string means use the specified path as the cache file.}

\item{config_init}{Additional configuration settings sent to
\code{\link[=POST]{POST()}}, e.g. \code{\link[=user_agent]{user_agent()}}.}

\item{credentials}{Advanced use only: allows you to completely customise
token generation.}

\item{validity_period}{A positive integer; token validity period in seconds.}

\item{insee_url}{Adresse du site fournissant les API. Ce paramètre n'est
utile que pour un usage interne à l'Insee, voir \link{insee_endpoint}.}
}
\value{
Un objet de classe \link{TokenInsee}.
}
\description{
Cette fonction génère un jeton d'accès à une application créée sur le site
\href{https://api.insee.fr/}{api.insee.fr}.
}
\keyword{internal}
