% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\docType{class}
\encoding{UTF-8}
\name{TokenInsee}
\alias{TokenInsee}
\title{Token objects for access to applications}
\format{Un objet de classe \code{R6}.}
\usage{
TokenInsee
}
\description{
Cette classe représente les jetons d'accès aux applications créées sur
\href{https://api.insee.fr}{api.insee.fr} et hérite de la classe
\link[httr:Token-class]{Token2.0} du package \link[httr:httr-package]{httr}. Les
objets de cette classe doivent être créés en appelant le constructeur
\code{get_insee_token()}.
}
\section{Methods}{

\itemize{
\item \code{cache()} : sauvegarde le jeton d'accès dans un cache
\item \code{revoke()} : révoque le jeton d'accès
}
}

\section{Caching}{


OAuth tokens are cached on disk in a file called \code{.httr-oauth}
saved in the current working directory.  Caching is enabled if:
\itemize{
\item The session is interactive, and the user agrees to it, OR
\item The \code{.httr-oauth} file is already present, OR
\item \code{getOption("httr_oauth_cache")} is \code{TRUE}
}

You can suppress caching by setting the \code{httr_oauth_cache} option to
\code{FALSE}.

Tokens are cached based on their endpoint and parameters.

The cache file should not be included in source code control or R packages
(because it contains private information), so httr will automatically add
the appropriate entries to \code{.gitignore} and \code{.Rbuildignore} if needed.

}

\keyword{internal}
