% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\docType{class}
\encoding{UTF-8}
\name{TokenInsee}
\alias{TokenInsee}
\title{Token objects for Insee applications}
\format{
Un objet de classe \code{R6}.
}
\description{
Cette classe représente les jetons d'accès aux applications créées sur
\href{https://api.insee.fr}{api.insee.fr} et hérite de la classe
\link[httr:Token-class]{Token2.0} du package \link[httr:httr-package]{httr}. Les
objets de cette classe doivent être créés en utilisant le constructeur
\link[=insee_token]{insee_token()}.
}
\section{Methods}{

\itemize{
\item \code{has_expired()} : le jeton d'accès a-t-il expiré ?
\item \code{cache()} : sauvegarde le jeton d'accès dans un cache.
\item \code{revoke()} : révoque le jeton d'accès.
\item \code{refresh()} : rafraichit le jeton d'accès (le point d'accès de
rafraichissement OAuth2 n'étant pas disponible, le jeton d'accès
courant est révoqué puis un nouveau jeton d'accès est généré).
}
}

\section{Caching}{


OAuth tokens are cached on disk in a file called \code{.httr-oauth}
saved in the current working directory.  Caching is enabled if:
\itemize{
\item The session is interactive, and the user agrees to it, OR
\item The \code{.httr-oauth} file is already present, OR
\item \code{getOption("httr_oauth_cache")} is \code{TRUE}
}

You can suppress caching by setting the \code{httr_oauth_cache} option to
\code{FALSE}.

Tokens are cached based on their endpoint and parameters.

The cache file should not be included in source code control or R packages
(because it contains private information), so httr will automatically add
the appropriate entries to \code{.gitignore} and \code{.Rbuildignore} if needed.

}

\keyword{internal}
\section{Super classes}{
\code{\link[httr:Token]{httr::Token}} -> \code{\link[httr:Token2.0]{httr::Token2.0}} -> \code{TokenInsee}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-print}{\code{TokenInsee$print()}}
\item \href{#method-init_credentials}{\code{TokenInsee$init_credentials()}}
\item \href{#method-has_expired}{\code{TokenInsee$has_expired()}}
\item \href{#method-load_from_cache}{\code{TokenInsee$load_from_cache()}}
\item \href{#method-revoke}{\code{TokenInsee$revoke()}}
\item \href{#method-refresh}{\code{TokenInsee$refresh()}}
\item \href{#method-can_refresh}{\code{TokenInsee$can_refresh()}}
\item \href{#method-clone}{\code{TokenInsee$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token" data-id="cache">}\href{../../httr/html/Token.html#method-cache}{\code{httr::Token$cache()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token" data-id="hash">}\href{../../httr/html/Token.html#method-hash}{\code{httr::Token$hash()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token" data-id="initialize">}\href{../../httr/html/Token.html#method-initialize}{\code{httr::Token$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token2.0" data-id="sign">}\href{../../httr/html/Token2.0.html#method-sign}{\code{httr::Token2.0$sign()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="httr" data-topic="Token2.0" data-id="validate">}\href{../../httr/html/Token2.0.html#method-validate}{\code{httr::Token2.0$validate()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init_credentials"></a>}}
\if{latex}{\out{\hypertarget{method-init_credentials}{}}}
\subsection{Method \code{init_credentials()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$init_credentials()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_expired"></a>}}
\if{latex}{\out{\hypertarget{method-has_expired}{}}}
\subsection{Method \code{has_expired()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$has_expired()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_from_cache"></a>}}
\if{latex}{\out{\hypertarget{method-load_from_cache}{}}}
\subsection{Method \code{load_from_cache()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$load_from_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-revoke"></a>}}
\if{latex}{\out{\hypertarget{method-revoke}{}}}
\subsection{Method \code{revoke()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$revoke()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-refresh}{}}}
\subsection{Method \code{refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-can_refresh"></a>}}
\if{latex}{\out{\hypertarget{method-can_refresh}{}}}
\subsection{Method \code{can_refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$can_refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenInsee$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
